+++ api/src/main/java/org/allaymc/api/command/CommandSender.java
@@ -3,12 +3,8 @@ package org.allaymc.api.command;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.i18n.TextReceiver;
-import org.allaymc.api.i18n.TrContainer;
 import org.allaymc.api.math.location.Location3dc;
 import org.allaymc.api.permission.Permissible;
-import org.allaymc.api.server.Server;
-import org.allaymc.api.world.gamerule.GameRule;
-import org.jetbrains.annotations.ApiStatus;
 
 /**
  * @author daoge_cmd
@@ -28,28 +24,6 @@ public interface CommandSender extends TextReceiver, Permissible {
      * @return the location where the command was executed
      */
     Location3dc getCommandExecuteLocation();
-
-    /**
-     * Handle the result of the command execution.
-     *
-     * @param result the result of the command execution
-     */
-    @ApiStatus.OverrideOnly
-    default void handleResult(CommandResult result) {
-        if (result.context() == null) return;
-        if (!getCommandExecuteLocation().dimension().getWorld().getWorldData().<Boolean>getGameRuleValue(GameRule.SEND_COMMAND_FEEDBACK)) {
-            return;
-        }
-
-        var status = result.status();
-        var outputs = result.context().getOutputs().toArray(TrContainer[]::new);
-        if (result.isSuccess()) {
-            Server.getInstance().broadcastCommandOutputs(result.context().getSender(), status, outputs);
-        } else {
-            // If there is an error, only send message to oneself
-            sendCommandOutputs(result.context().getSender(), status, outputs);
-        }
-    }
 
     /**
      * Checks if the sender is a player.